examples/requirements.txt
# Core RAG Dependencies
chromadb>=0.4.0
sentence-transformers>=2.2.0
openai>=1.0.0
python-dotenv>=1.0.0

# Additional useful packages
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
flask>=2.3.0
requests>=2.31.0

# Development dependencies
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: Alternative vector databases
# pinecone-client>=2.2.0
# weaviate-client>=3.20.0

# Optional: Alternative embedding models
# transformers>=4.30.0
# torch>=2.0.0